# swagger-ui/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  Chain:
    description: Chain represents the necessary data for connecting to and indentifying
      a chain and its counterparites
    properties:
      account-prefix:
        type: string
        x-go-name: AccountPrefix
      chain-id:
        type: string
        x-go-name: ChainID
      gas-adjustment:
        format: double
        type: number
        x-go-name: GasAdjustment
      gas-prices:
        type: string
        x-go-name: GasPrices
      key:
        type: string
        x-go-name: Key
      rpc-addr:
        type: string
        x-go-name: RPCAddr
      trusting-period:
        type: string
        x-go-name: TrustingPeriod
    type: object
    x-go-package: github.com/cosmos/relayer/relayer
...
  addChainRequest:
    properties:
      account-prefix:
        type: string
        x-go-name: AccountPrefix
      file:
        type: string
        x-go-name: FilePath
      gas-adjustment:
        type: string
        x-go-name: GasAdjustment
      gas-prices:
        type: string
        x-go-name: GasPrices
      key:
        type: string
        x-go-name: Key
      rpc-addr:
        type: string
        x-go-name: RPCAddr
      trusting-period:
        type: string
        x-go-name: TrustingPeriod
      url:
        type: string
        x-go-name: URL
    type: object
    x-go-package: github.com/cosmos/relayer/cmd
...
host: localhost:5183
info:
  description: A REST interface for state queries.
  title: Relayer Rest Server.
  version: 1.0.0
paths:
...
  /chains/{name}:
    get:
      operationId: getChain
      parameters:
      - in: path
        name: name
        required: true
        type: string
        x-go-name: Name
      responses:
        "200":
          $ref: '#/responses/getChainResponse'
        "400":
          $ref: '#/responses/errorResponse'
      summary: Get single chain details.
      tags:
      - Chains
    post:
      description: Add a chain
      operationId: addChain
      parameters:
      - in: path
        name: name
        required: true
        type: string
        x-go-name: Name
      - in: body
        name: Body
        schema:
          $ref: '#/definitions/addChainRequest'
      responses:
        "201":
          $ref: '#/responses/addChainResponse'
        "400":
          $ref: '#/responses/errorResponse'
        "500":
          $ref: '#/responses/errorResponse'
      tags:
      - Chains
...
produces:
- application/json
responses:
  addChainResponse:
    description: ""
...
  errorResponse:
    description: ""
    schema:
      properties:
        err:
          type: string
          x-go-name: Err
      type: object
  getChainResponse:
    description: ""
    schema:
      $ref: '#/definitions/Chain'
...
schemes:
- http
swagger: "2.0"